Recurate changeux2012 (#167)
* Fixing errors in handling chnageux-2012

* Recurate-chenguex 2012-trusha19

I guess due to some copy-paste error, the paper is duplicated in the document. Amongst both, the upper copy is recurated.

There were some ambiguous evidences which are marked with low confidence level. I would like to know a possible solution of those.

* Recurate chenguex2012- trusha19

The copied part of the paper was removed

* Recurate-changeux2012-trusha19

* Review changeux2012

* Update changeux2012.bel

* Update changeux2012.bel